now all package files have coverage >= 90%
